+++ src/main/java/com/yechtech/dac/traceability/domain/BatchTraceabilityLc.java
@@ -1,6 +1,7 @@
 package com.yechtech.dac.traceability.domain;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Date;
 
 import com.baomidou.mybatisplus.annotation.TableName;
@@ -51,11 +52,11 @@ public class BatchTraceabilityLc extends Model<BatchTraceabilityLc> implements S
     @ApiModelProperty(notes = "最晚出货时间")
     private Date latestshipdate;
     @ApiModelProperty(notes = "到货数量")
-    private Integer receivedQty;
+    private BigDecimal receivedQty;
     @ApiModelProperty(notes = "出货数量")
-    private Integer shipmentQty;
+    private BigDecimal shipmentQty;
     @ApiModelProperty(notes = "在库数量")
-    private Integer inventoriesQty;
+    private BigDecimal inventoriesQty;
     @ApiModelProperty(notes = "采购单位")
     private String uom;
     @ApiModelProperty(notes = "psid")
